pyproject.toml


[project]
name = "interview-system"
version = "0.1.0"
description = "Multi-Agentic RAG Interview System"
authors = [
  { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.11"
dependencies = [
  # --- Core Frameworks ---
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.29.0",

  # --- AI & LangChain (Course Versions) ---
  "langchain-anthropic>=0.3.10",
  "langchain-community>=0.3.20",
  "langchain-experimental>=0.3.4",
  "langchain-openai>=0.3.9",
  "langgraph>=0.3.18",
  "langgraph-checkpoint-sqlite>=2.0.6",
  "langsmith>=0.3.18",
  "langchain-google-genai>=1.0.0",

  # --- Database & ORM ---
  "sqlalchemy[asyncio]>=2.0.30",
  "alembic>=1.13.1",
  "asyncpg>=0.29.0",
  "psycopg2-binary>=2.9.0",
  "cloudinary >= 1.39.0",
  "python-multipart>=0.0.9",

  # --- Vector DB & Cache ---
  #"chromadb>=0.5.0",
  #"redis>=5.0.4",

  # --- Auth & Security ---
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",

  # --- Utilities & Config ---
  "pydantic[email]>=2.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "jinja2>=3.1.4",
  "pypdf >= 4.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.23.6",
  "httpx>=0.27.0",
  "mypy>=1.10.0",
  "pre-commit>=3.7.1",
]

# Tool configurations remain the same
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
python_files = "test_*.py"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]
